+++ public/js/player.js
@@ -1,0 +1,27 @@
+(function (){
+    'use strict';
+
+    var noSleep;
+
+    var toggleSleepCtrl = document.getElementById('prevent-sleep');
+
+    toggleSleepCtrl.addEventListener('click', handleClick);
+
+    function handleClick(event) {
+        debugger;
+        if(!noSleep){
+            if(!NoSleep){
+                event.preventDefault();
+                return;
+            }
+
+            noSleep = new NoSleep();
+        }
+
+        if(event.target.checked){
+            noSleep.enable();
+        }else{
+            noSleep.disable();
+        }
+    }
+})();